requirements.txt
asgiref==3.8.1
bcrypt==4.0.1
certifi==2024.7.4
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.1
Django==4.2.5
django-autocomplete-light==3.9.7
django-braces==1.15.0
django-cleanup==8.0.0
django-cors-headers==4.4.0
django-localflavor==4.0
django-multiselectfield==0.1.12
django-recaptcha==3.0.0
django-rest-swagger==2.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
idna==3.7
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
openapi-codec==1.3.2
packaging==24.1
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
python-stdnum==1.20
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
simplejson==3.19.2
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.9
